runner: add Network Block Device mapping tool support
Currently, this is based the glusterfs and will only support the loop
back nbd device locally. And it will map the volume file to one nbd device.

This patch is based the nbd netlink feature and need to use the
upstream kernel and in theory the nbd device number is unlimited.

Usage:
	nbd <command> [<args>]

Commands:
	help
		display help for nbd commands

	create <volname@host:/path> [prealloc <full|no>] <size SIZE>
		create path file on the volname volume, prealloc is no as default,
		and the SIZE is valid with B, K(iB), M(iB), G(iB), T(iB), P(iB), E(iB), Z(iB), Y(iB)

	delete <volname@host:/path>
		delete path file on the volname volume

	map <volname@host:/path> [nbd-device] [threads NUM] [timeout TIME] [daemon on|off]
		map path file to the nbd device, as default the threads 4, timeout 0 and daemon on

	umap <nbd-device>
		umap the nbd device

	list <mapd|umap|all>
		list the mapped|umapped|all nbd devices, all as default

	version
		show version info and exit.

Next will make one daemon to handle all the mapping request from
the client side.

Signed-off-by: Xiubo Li <xiubli@redhat.com>