runner: Add nbd-runner and nbd-cli support and switch to autoconfig
From this commit, there will be one Server daemon nbd-runner running,
which mostly will run together with the backend storage service, such
as glusterfs. And the nbd-cli cmd could run any whether as you like.

nbd-cli is using the RPC to control and send the commands to nbd-runner,
and the I/Os will in another socket path.

TODO:
1, add logger support
2, split the gluster code and make it as one handler, which could more
   easy to add other handler support.
3, add read-only options support.
4, add UNIX support if the client and server are both in the same node.
5, split the nbd-cli and make it more easy to add other handler support.

Signed-off-by: Xiubo Li <xiubli@redhat.com>