cli: set nbds_max to 0 when insearting the nbd.ko
This is mainly to avoid the following issue:

Currently in nbd.ko module for the first time when loading, it will
add and initialize nbds_max(16 as default) nbd devices defautly, and
the udev service will try to open/close the /dev/nbdX to do some sanity
check.

For the first time to execute the 'nbd-cli ... list' command just after
the nbd.ko is inserted we will randomly hit some unused /dev/nbdX devices
are listed as mapped, due to nbd kernel module bug.

If there really needs to set nbds_max to other number instead of 0 here,
we should uncomment the "sleep(1);".

Signed-off-by: Xiubo Li <xiubli@redhat.com>