clid: restore the stale connections after node/runner rebooted
TODO:
There is one problem that when trying to check the nbd device
NBD_CMD_STATUS and at the same time insert the nbd.ko module,
we can randomly get some of the 16 /dev/nbd{0~15} are connected,
but they are not. This is because that the udev service in user
space will try to open /dev/nbd{0~15} devices to do some sanity
check when they are added in "__init nbd_init()" and then close
it asynchronousely.

And the NBD_CMD_DISCONNECT still has the similiar problem, so
we need to wait for a while.

This should be fixed in kernel space. And here as one work
around we just hard code it and wait 1 seconds.

Signed-off-by: Xiubo Li <xiubli@redhat.com>