azblk: Recover from Azure tcp RST (#86)
When running IO to a blob page we see Azure reset the connection multiple
times (CURLE_SEND_ERROR/CURLE_RECV_ERROR). This can happen frequently on
some hosts and not at all on others. MS has not yet analyzed the problem.
Since it seems to be a recoverable error azblk will issue a retry of the
request until success or command timeout.

Signed-off-by: Cathy Avery <cavery@redhat.com>